+++ src/irvine/oeis/a034/A034734.java
@@ -18,7 +18,7 @@ public class A034734 extends Sequence1 {
     Z sum = Z.ZERO;
     for (final Z dd : Jaguar.factor(++mN).divisors()) {
       final int d = dd.intValue();
-      sum = sum.add(Functions.FIBONACCI.z((long) (mN / d)).shiftLeft(d - 1));
+      sum = sum.add(Functions.FIBONACCI.z(mN / d).shiftLeft(d - 1));
     }
     return sum;
   }
