+++ src/irvine/oeis/a045/A045662.java
@@ -22,7 +22,7 @@ public class A045662 extends Sequence0 {
     Z sum = Z.ZERO;
     for (final Z dd : Jaguar.factor(mN).divisors()) {
       final int d = dd.intValue();
-      sum = sum.add(Binomial.binomial(d & ~1, d / 2).multiply(Functions.MOBIUS.i((long) (mN / d))));
+      sum = sum.add(Binomial.binomial(d & ~1, d / 2).multiply(Functions.MOBIUS.i(mN / d)));
     }
     return sum.multiply(2L * mN);
   }
